+++ ui/box/CHANGELOG.md
@@ -3,6 +3,24 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.21.0](https://github.com/washingtonpost/wpds-ui-kit/compare/v1.20.0...v1.21.0) (2024-02-07)
+
+**Note:** Version bump only for package @washingtonpost/wpds-box
+
+# [1.20.0](https://github.com/washingtonpost/wpds-ui-kit/compare/v1.19.0...v1.20.0) (2024-01-24)
+
+**Note:** Version bump only for package @washingtonpost/wpds-box
+
+# [1.19.0](https://github.com/washingtonpost/wpds-ui-kit/compare/v1.18.0...v1.19.0) (2024-01-10)
+
+**Note:** Version bump only for package @washingtonpost/wpds-box
+
+# [1.18.0](https://github.com/washingtonpost/wpds-ui-kit/compare/v1.17.0...v1.18.0) (2023-12-13)
+
+### Bug Fixes
+
+- our lockfile area for build dot site was borked ([#547](https://github.com/washingtonpost/wpds-ui-kit/issues/547)) ([2a66631](https://github.com/washingtonpost/wpds-ui-kit/commit/2a66631d066b7128c4be209a60af0ad86fc528e9))
+
 # [1.17.0](https://github.com/washingtonpost/wpds-ui-kit/compare/v1.16.2...v1.17.0) (2023-11-30)
 
 **Note:** Version bump only for package @washingtonpost/wpds-box